+++ Changelog.md
@@ -10,6 +10,12 @@ This file present the software status in form of a "Changelog".
 
 This document is valid within the scope of the work for all projects.
 
+## [Unreleased]
+
+### Updated
+
+* Update Result pattern to change error to an interface and generate a generic Error class.
+
 ## 2.13.2
 
 ### Fixed